🚀 New Features
- Title ratings: Added ability to rate titles
- Username change: Implemented username modification functionality
- API synchronization: Updated methods for latest shinga-api version

🐛 Bug Fixes
- Scroll performance: Fixed infinite scroll issue during fast scrolling
- Favorites display: Resolved incorrect titles rendering on favorites updates
- UI improvements: Various minor UI fixes and enhancements

🔧 Technical Improvements
- Cache system: Overhauled cache display management in settings
- Database storage: Changed local database files directory structure
- App icon: Added application icon assets

📦 Version Update
- Bumped application version to 0.9.0

🔄 Breaking Changes
- Requires updated shinga-api backend
- Database storage location changed

+++ lib/data/models/title/title.dart
@@ -41,7 +41,7 @@ abstract class TitleData with _$TitleData {
     required int chapters,
     required int volumes,
     required int views,
-    required int inAppRating,
+    required double inAppRating,
     required int inAppScoredBy,
     @TitleStatusConverter() required TitleStatus status,
     required TitleReleaseTime date,
